separate supervisor-lib from common-lib for items that should not be exposed to apps
