+++ services/user/CommonSys/common/packages/common-lib/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./keyConversions";
 export * from "./rpc";
 export * from "./messaging";
+export { Supervisor } from "./supervisor";
