+++ services/user/SupervisorSys/lib/.prettierrc.json
@@ -1,0 +1,4 @@
+{
+    "tabWidth": 4,
+    "plugins": ["prettier-plugin-tailwindcss"]
+}
