+++ services/user/SupervisorSys/lib/package.json
@@ -1,0 +1,36 @@
+{
+    "name": "@psibase/supervisor-lib",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "main": "./dist/supervisor-lib.umd.cjs",
+    "module": "./dist/supervisor-lib.js",
+    "types": "./dist/index.d.ts",
+    "exports": {
+        ".": {
+            "import": "./dist/supervisor-lib.js",
+            "require": "./dist/supervisor-lib.umd.cjs"
+        }
+    },
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "dev": "vite",
+        "build": "tsc && vite build",
+        "preview": "vite preview",
+        "format": "prettier --write 'src/**/*.ts'"
+    },
+    "devDependencies": {
+        "@types/elliptic": "^6.4.14",
+        "@types/node": "^18.15.0",
+        "prettier": "^3.2.5",
+        "prettier-plugin-tailwindcss": "^0.5.13",
+        "typescript": "^4.9.3",
+        "vite": "^4.1.0"
+    },
+    "dependencies": {
+        "elliptic": "^6.5.4",
+        "hash.js": "^1.1.7"
+    }
+}
