+++ services/user/SupervisorSys/ui/tsconfig.json
@@ -1,30 +1,34 @@
 {
-  "compilerOptions": {
-    "target": "ESNext",
-    "useDefineForClassFields": true,
-    "module": "ESNext",
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
-    "skipLibCheck": true,
+    "compilerOptions": {
+        "target": "ESNext",
+        "useDefineForClassFields": true,
+        "module": "ESNext",
+        "lib": ["ES2020", "DOM", "DOM.Iterable"],
+        "skipLibCheck": true,
 
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
+        /* Bundler mode */
+        "moduleResolution": "bundler",
+        "allowImportingTsExtensions": true,
+        "resolveJsonModule": true,
+        "isolatedModules": true,
+        "noEmit": true,
 
-    "paths": {
-      "@psibase/common-lib": ["../../CommonSys/common/packages/common-lib/src"],
-      "@psibase/common-lib/*": [
-        "../../CommonSys/common/packages/common-lib/src/*"
-      ]
+        "paths": {
+            "@psibase/common-lib": [
+                "../../CommonSys/common/packages/common-lib/src"
+            ],
+            "@psibase/common-lib/*": [
+                "../../CommonSys/common/packages/common-lib/src/*"
+            ],
+            "@psibase/supervisor-lib": ["../lib/src"],
+            "@psibase/supervisor-lib/*": ["../lib/src/*"]
+        },
+
+        /* Linting */
+        "strict": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
+        "noFallthroughCasesInSwitch": true
     },
-
-    /* Linting */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "include": ["src"]
+    "include": ["src"]
 }
